requirements.txt
google-generativeai>=0.7.0
pypdf>=4.2.0
python-dotenv>=1.0.1
gradio>=4.44.0
